+++ api/yq/assetvk/rendered/Diamond3.cpp
@@ -124,7 +124,7 @@ namespace yq::tachyon {
 
 
         {
-            auto& p = w.pipeline(Pipeline::Role::SolidColor);
+            auto& p = w.pipeline(pipekey::SOLID_COLOR);
             
             p.shader("yq/shape3/color.vert");
             p.shader("yq/shape3/color.frag");
@@ -139,7 +139,7 @@ namespace yq::tachyon {
         }
 
         {
-            auto& p = w.pipeline(Pipeline::Role::ColorCorner);
+            auto& p = w.pipeline(pipekey::CORNER_COLOR);
             
             p.shader("yq/shape3/gradient.vert");
             p.shader("yq/shape3/gradient.frag");
@@ -185,7 +185,7 @@ namespace yq::tachyon {
         switch(draw_mode(USE)){
         case DrawMode::Color:
             m_good      = true;
-            set_pipeline(Pipeline::Role::SolidColor);
+            set_pipeline(pipekey::SOLID_COLOR);
             m_vertexS   = {
                 vs(vertex_east()), vs(vertex_west()), vs(vertex_north()), vs(vertex_south()), 
                 vs(vertex_top()), vs(vertex_bottom())
@@ -196,7 +196,7 @@ namespace yq::tachyon {
             break;
         case DrawMode::Gradient:
             m_good      = true;
-            set_pipeline(Pipeline::Role::ColorCorner);
+            set_pipeline(pipekey::CORNER_COLOR);
             m_vertexC = {
                 vc(vertex_east()), vc(vertex_west()), vc(vertex_north()), vc(vertex_south()), 
                 vc(vertex_top()), vc(vertex_bottom())