+++ api/yq/assetvk/rendered/Tetrahedron3.cpp
@@ -127,7 +127,7 @@ namespace yq::tachyon {
         w.property("v4", &Tetrahedron³::v4);
 
         {
-            auto& p = w.pipeline(Pipeline::Role::SolidColor);
+            auto& p = w.pipeline(pipekey::SOLID_COLOR);
             
             p.shader("yq/shape3/color.vert");
             p.shader("yq/shape3/color.frag");
@@ -142,7 +142,7 @@ namespace yq::tachyon {
         }
 
         {
-            auto& p = w.pipeline(Pipeline::Role::ColorCorner);
+            auto& p = w.pipeline(pipekey::CORNER_COLOR);
             
             p.shader("yq/shape3/gradient.vert");
             p.shader("yq/shape3/gradient.frag");
@@ -196,7 +196,7 @@ namespace yq::tachyon {
         switch(draw_mode(USE)){
         case DrawMode::Color:
             m_good      = true;
-            set_pipeline(Pipeline::Role::SolidColor);
+            set_pipeline(pipekey::SOLID_COLOR);
             m_vertexS   = {
                 vs(vertex1()), vs(vertex2()), vs(vertex3()), vs(vertex4())
             };
@@ -206,7 +206,7 @@ namespace yq::tachyon {
             break;
         case DrawMode::Gradient:
             m_good      = true;
-            set_pipeline(Pipeline::Role::ColorCorner);
+            set_pipeline(pipekey::CORNER_COLOR);
             m_vertexC = {
                 vc(vertex1()), vc(vertex2()), vc(vertex3()), vc(vertex4())
             };
@@ -221,7 +221,7 @@ namespace yq::tachyon {
 #if 0
     void    Tetrahedron³::rebuild_textured()
     {
-        set_pipeline(Pipeline::Role::Textured);
+        set_pipeline(PipelineKey::Textured);
         m_vertexT = {
             vt(vertex1()), vt(vertex2()), vt(vertex3()), vt(vertex4())
         };